# services/nvs-runtime/runtime/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NVS MCP Runtime Server",
    "description": "RFC-NVS42 v0.3 \u2014 Agent Registry, Session Management, SEP Event Transport",
    "version": "0.1.0-p1"
  },
  "paths": {
    "/api/v1/agents": {
      "get": {
        "summary": "List Agents",
        "operationId": "list_agents_api_v1_agents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentResponse"
                  },
                  "type": "array",
                  "title": "Response List Agents Api V1 Agents Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Agent",
        "operationId": "create_agent_api_v1_agents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}": {
      "get": {
        "summary": "Get Agent",
        "operationId": "get_agent_api_v1_agents__agent_id__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions": {
      "post": {
        "summary": "Create Session",
        "operationId": "create_session_api_v1_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}": {
      "get": {
        "summary": "Get Session",
        "operationId": "get_session_api_v1_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/join": {
      "post": {
        "summary": "Join Session",
        "operationId": "join_session_api_v1_sessions__session_id__join_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionJoin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/experiments": {
      "post": {
        "summary": "Create Experiment",
        "operationId": "create_experiment_api_v1_sessions__session_id__experiments_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/experiments/{experiment_id}": {
      "get": {
        "summary": "Get Experiment",
        "operationId": "get_experiment_api_v1_experiments__experiment_id__get",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/events": {
      "post": {
        "summary": "Ingest Event",
        "operationId": "ingest_event_api_v1_sessions__session_id__events_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventIngestRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventIngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Query Events",
        "operationId": "query_events_api_v1_sessions__session_id__events_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventResponse"
                  },
                  "title": "Response Query Events Api V1 Sessions  Session Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/stream": {
      "get": {
        "summary": "Stream Events",
        "operationId": "stream_events_api_v1_sessions__session_id__stream_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "replay_from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0-0",
              "title": "Replay From"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/capabilities": {
      "get": {
        "summary": "Capabilities",
        "operationId": "capabilities_api_v1_capabilities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "summary": "Ready",
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentCreate": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "provider": {
            "$ref": "#/components/schemas/AgentProvider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "transport": {
            "type": "string",
            "title": "Transport",
            "default": "mcp"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          },
          "metadata": {
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "model"
        ],
        "title": "AgentCreate"
      },
      "AgentProvider": {
        "type": "string",
        "enum": [
          "anthropic",
          "openai",
          "google",
          "custom"
        ],
        "title": "AgentProvider"
      },
      "AgentResponse": {
        "properties": {
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "transport": {
            "type": "string",
            "title": "Transport"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "metadata": {
            "type": "object",
            "title": "Metadata"
          },
          "registered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Registered At"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "display_name",
          "provider",
          "model",
          "transport",
          "capabilities",
          "status",
          "metadata",
          "registered_at",
          "last_seen_at"
        ],
        "title": "AgentResponse"
      },
      "CapabilitiesResponse": {
        "properties": {
          "service": {
            "type": "string",
            "title": "Service"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "rfc": {
            "type": "string",
            "title": "Rfc"
          },
          "features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Features"
          },
          "mcp_tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mcp Tools"
          },
          "layer3_stubs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Layer3 Stubs"
          }
        },
        "type": "object",
        "required": [
          "service",
          "version",
          "rfc",
          "features",
          "mcp_tools",
          "layer3_stubs"
        ],
        "title": "CapabilitiesResponse"
      },
      "EventIngestRequest": {
        "properties": {
          "event_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeEventType"
              },
              {
                "type": "string"
              }
            ],
            "title": "Event Type"
          },
          "experiment_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Id"
          },
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Id"
          },
          "source_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Provider"
          },
          "source_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Model"
          },
          "parent_event_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Event Id"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "payload": {
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "event_type",
          "agent_id"
        ],
        "title": "EventIngestRequest"
      },
      "EventIngestResponse": {
        "properties": {
          "accepted": {
            "type": "integer",
            "title": "Accepted"
          },
          "event_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Event Ids"
          },
          "forward_status": {
            "$ref": "#/components/schemas/ForwardStatus"
          }
        },
        "type": "object",
        "required": [
          "accepted",
          "event_ids",
          "forward_status"
        ],
        "title": "EventIngestResponse"
      },
      "EventResponse": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "experiment_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "agent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "source_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Provider"
          },
          "source_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Model"
          },
          "parent_event_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Event Id"
          },
          "sequence_id": {
            "type": "integer",
            "title": "Sequence Id"
          },
          "sep_event_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sep Event Type"
          },
          "payload": {
            "type": "object",
            "title": "Payload"
          },
          "forward_status": {
            "type": "string",
            "title": "Forward Status"
          },
          "kernel_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kernel Run Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "schema_version",
          "event_type",
          "experiment_id",
          "session_id",
          "agent_id",
          "source_provider",
          "source_model",
          "parent_event_id",
          "sequence_id",
          "sep_event_type",
          "payload",
          "forward_status",
          "kernel_run_id",
          "created_at"
        ],
        "title": "EventResponse"
      },
      "ExperimentCreate": {
        "properties": {
          "experiment_type": {
            "$ref": "#/components/schemas/ExperimentType",
            "default": "sep"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "metadata": {
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "title": "ExperimentCreate"
      },
      "ExperimentResponse": {
        "properties": {
          "experiment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Experiment Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "experiment_type": {
            "type": "string",
            "title": "Experiment Type"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "metadata": {
            "type": "object",
            "title": "Metadata"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          }
        },
        "type": "object",
        "required": [
          "experiment_id",
          "session_id",
          "experiment_type",
          "label",
          "status",
          "metadata",
          "started_at",
          "ended_at"
        ],
        "title": "ExperimentResponse"
      },
      "ExperimentType": {
        "type": "string",
        "enum": [
          "sep",
          "sysid",
          "control",
          "custom"
        ],
        "title": "ExperimentType"
      },
      "ForwardStatus": {
        "type": "string",
        "enum": [
          "pending",
          "queued",
          "forwarded",
          "skipped",
          "failed"
        ],
        "title": "ForwardStatus"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RuntimeEventType": {
        "type": "string",
        "enum": [
          "session.join",
          "session.leave",
          "narrative.append",
          "state.raw",
          "trace.fragment",
          "sep.excitation",
          "sep.disturbance",
          "sep.control",
          "sep.raw",
          "heartbeat",
          "error"
        ],
        "title": "RuntimeEventType"
      },
      "SessionCreate": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "participants": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Participants"
          },
          "options": {
            "type": "object",
            "title": "Options"
          }
        },
        "type": "object",
        "title": "SessionCreate"
      },
      "SessionJoin": {
        "properties": {
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Id"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "participant"
          }
        },
        "type": "object",
        "required": [
          "agent_id"
        ],
        "title": "SessionJoin"
      },
      "SessionResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "kernel_run_ids": {
            "type": "object",
            "title": "Kernel Run Ids"
          },
          "options": {
            "type": "object",
            "title": "Options"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "participants": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Participants"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "label",
          "status",
          "kernel_run_ids",
          "options",
          "created_at",
          "closed_at"
        ],
        "title": "SessionResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}